Move some modules around
