--- librad/src/internal.rs
@@ -1,1 +1,0 @@
-pub(crate) mod channel;
