--- librad/src/util.rs
@@ -1,1 +1,0 @@
-pub mod monitor;
